requirements.txt
fastapi>=0.98.0
pandas>=2.0.3
starlette>=0.27.0
uvicorn>=0.22.0 
structlog>=23.1.0
requests~=2.32.3
aiohttp~=3.9.5
python-dateutil~=2.9.0.post0
python-dotenv~=1.0.1
numpy~=2.0.0
pytest~=8.2.2